Dynamic threshold (#40)
When detecting sproket hole template, if the match level returned by OpenVision is less than 0.9, additional retries will be done (decreasing the threshold level by 10 each retry) until reaching a match level of 0.9 or higher. If a threshold level of 50 is reached, then the best match level found so far is retained. With this change, I have been able to reduce the percentage of misaligned frames from 4.2% to 1.4%. The drawback is that it takes a bit longer (due to additional retries for badly matched frames) but the end result is worth the wait.